Add progress bar
The creation of the generators (see main function in object_detector_retinanet/keras_retinanet/bin/train.py)
takes several minutes.
We do not have any feedback when running the trainning script which
can be misleading. Most of the time is spent in the loop in _read_images() in the file
object_detector_retinanet/keras_retinanet/preprocessing/csv_generator.py so I added a progress bar with
tqdm (https://tqdm.github.io/). Now when we run the training script we know what the script is doing and
how much time we need to wait for the generator to complete.