bugfix: issue #2 TypeError in preprocessing
The error raised (see. issue #2) comes from the fact that annotation
coordinates are read from a file, hence they are typed as str.
When compared with an int value, the TypeError is raised since
it is not allowed to compare int and str with comparison operators.
A cast from str to int solves the problem.