feat: Divide time to load images by 1500 🍾🍾
Before: images were load in memory to get their dimensions.
Now: just read the header of the file as the dimensions are specified in it.

Result: Time to load images reduces by a factor of 1500-2000! This is particularly
appreciable since it takes more than 70min every time I launch the train.