fix: trim box coordinates
Some of our annotations contains small errors. In fact
the bounding box coordinates are slightly out of the image dimensions.

For instance let consider an image of size 600x600, we have some annoations denoting
object positions that can end in posX=601, instead of throwing the entire image, we resize
the annotation so that it never ends outside the image.